refactor: move vscode utilities to src/frontend/vscode/
Move vscodeDiagnostics.ts and vscodeEditor.ts to src/frontend/vscode/
since these are VS Code extension-host specific utilities.
