+++ src/frontend/vscode/vscodeDiagnostics.ts
